# backend/internal/handler/v1/swagger/swagger.yaml
definitions:
...
  dd-prediction-api_internal_model.SolanaToken:
    properties:
      decimals:
        type: integer
      image_url:
        type: string
      is_verified:
        type: boolean
      market_cap:
        type: number
      mint:
        type: string
      name:
        type: string
      program_id:
        type: string
      symbol:
        type: string
      updated_at:
        type: string
      usd_price:
        type: number
    type: object
...
  dd-prediction-api_pkg_apperrors.ErrorPublic:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a swagger specification for a Duel Duck back-end.
  title: Duel Duck Prediction API
  version: "1.0"
paths:
...
  /coins/mint/{mint}:
    get:
      consumes:
      - application/json
      description: Retrieve Solana token information by searching for tokens with
        the specified mint address
      parameters:
      - description: Mint address to search for
        in: path
        name: mint
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token information retrieved successfully
          schema:
            $ref: '#/definitions/dd-prediction-api_internal_model.SolanaToken'
        "400":
          description: Invalid request data or invalid mint address format
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "404":
          description: Token with provided mint address not found
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error during token retrieval
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      summary: Get Solana token by mint address
      tags:
      - coins
  /coins/name/{name}:
    get:
      consumes:
      - application/json
      description: Retrieve Solana token information by searching for tokens with
        the specified name
      parameters:
      - description: Token name to search for
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of tokens matching the name
          schema:
            items:
              $ref: '#/definitions/dd-prediction-api_internal_model.SolanaToken'
            type: array
        "400":
          description: Invalid request data or empty token name
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error during token retrieval
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      summary: Get Solana token by name
      tags:
      - coins
  /coins/solana-tokens:
    get:
      consumes:
      - application/json
      description: Search Solana tokens by symbol or name (query params). Returns
        list of matching tokens.
      parameters:
      - description: Search string (symbol or name)
        in: query
        name: search
        required: true
        type: string
      - default: 20
        description: Max results to return
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of matching tokens
          schema:
            items:
              $ref: '#/definitions/dd-prediction-api_internal_model.SolanaToken'
            type: array
        "400":
          description: Invalid request params
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      summary: Search Solana tokens
      tags:
      - coins
  /coins/solana-tokens/price:
    get:
      consumes:
      - application/json
      description: Retrieve current token price by mint address (query param).
      parameters:
      - description: Token mint address
        in: query
        name: mint
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token price retrieved successfully
          schema:
            properties:
              price:
                type: number
            type: object
        "400":
          description: Invalid request params
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "404":
          description: Token not found
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
        "503":
          description: External service unavailable
          schema:
            $ref: '#/definitions/dd-prediction-api_pkg_apperrors.ErrorPublic'
      summary: Get Solana token price
      tags:
      - coins
...
securityDefinitions:
  BearerAuth:
    description: Enter the auth token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"